feat(product): add favorite button on product card
