+++ @ecomplus/storefront-app/src/lib/i18n.js
@@ -502,6 +502,13 @@ export const ZipCode = {
   pt_br: 'CEP'
 }
 
+/*
+  export const AddToFavorites = {
+  en_us: 'Add to Favorites',
+  pt_br: 'Adicionar aos favoritos'
+}
+*/
+
 export const _FinancialStatus = {
   pending: {
     en_us: 'Pending payment',
